moved og defaults to view layer
